+++ src/Components/Common/dashboardWrapper.css
@@ -301,7 +301,7 @@ body.dark .switch::before {
     color: var(--text-color);
 } */
 
-/* ==========================DashboardWrapper2 starts=======================  */
+/* ==========================DashboardWrapper starts=======================  */
 
 .s_nav_container {
   display: inline-flex;
